Improvements noticed while reviewing the changes
