+++ diesel/src/pg/types/floats/mod.rs
@@ -118,6 +118,11 @@ impl FromSql<sql_types::Float, Pg> for f32 {
             "Received more than 4 bytes while decoding \
              an f32. Was a double accidentally marked as float?"
         );
+        debug_assert!(
+            bytes.len() >= 4,
+            "Received less than 4 bytes while decoding \
+             an f32."
+        );
         bytes
             .read_f32::<NetworkEndian>()
             .map_err(|e| Box::new(e) as Box<dyn Error + Send + Sync>)
@@ -129,6 +134,11 @@ impl FromSql<sql_types::Double, Pg> for f64 {
         let mut bytes = value.as_bytes();
         debug_assert!(
             bytes.len() <= 8,
+            "Received less than 8 bytes while decoding \
+             an f64. Was a float accidentally marked as double?"
+        );
+        debug_assert!(
+            bytes.len() >= 8,
             "Received more than 8 bytes while decoding \
              an f64. Was a numeric accidentally marked as double?"
         );
