+++ diesel/src/r2d2.rs
@@ -130,14 +130,14 @@ where
     }
 }
 
-impl<'a, 'b, DB, M> ConnectionGatWorkaround<'a, 'b, DB> for PooledConnection<M>
+impl<'conn, 'query, DB, M> ConnectionGatWorkaround<'conn, 'query, DB> for PooledConnection<M>
 where
     M: ManageConnection,
     M::Connection: Connection<Backend = DB>,
     DB: Backend,
 {
-    type Cursor = <M::Connection as ConnectionGatWorkaround<'a, 'b, DB>>::Cursor;
-    type Row = <M::Connection as ConnectionGatWorkaround<'a, 'b, DB>>::Row;
+    type Cursor = <M::Connection as ConnectionGatWorkaround<'conn, 'query, DB>>::Cursor;
+    type Row = <M::Connection as ConnectionGatWorkaround<'conn, 'query, DB>>::Row;
 }
 
 impl<M> Connection for PooledConnection<M>
@@ -159,13 +159,13 @@ where
         (&mut **self).execute(query)
     }
 
-    fn load<'a, 'b, T>(
-        &'a mut self,
+    fn load<'conn, 'query, T>(
+        &'conn mut self,
         source: T,
-    ) -> QueryResult<<Self as ConnectionGatWorkaround<'a, 'b, Self::Backend>>::Cursor>
+    ) -> QueryResult<<Self as ConnectionGatWorkaround<'conn, 'query, Self::Backend>>::Cursor>
     where
         T: AsQuery,
-        T::Query: QueryFragment<Self::Backend> + QueryId + 'b,
+        T::Query: QueryFragment<Self::Backend> + QueryId + 'query,
         Self::Backend: QueryMetadata<T::SqlType>,
     {
         (&mut **self).load(source)
