+++ diesel/src/sqlite/connection/statement_iterator.rs
@@ -19,7 +19,7 @@ enum PrivateStatementIterator<'stmt, 'query> {
 }
 
 impl<'stmt, 'query> StatementIterator<'stmt, 'query> {
-    pub fn new(stmt: StatementUse<'stmt, 'query>) -> Self {
+    pub fn new(stmt: StatementUse<'stmt, 'query>) -> StatementUse<'stmt, 'query> {
         Self {
             inner: PrivateStatementIterator::NotStarted(stmt),
             column_names: None,
