+++ diesel/src/sqlite/mod.rs
@@ -15,6 +15,7 @@ pub use self::backend::{Sqlite, SqliteType};
 pub use self::connection::SqliteConnection;
 pub use self::connection::SqliteValue;
 pub use self::query_builder::SqliteQueryBuilder;
+pub(crate) use self::connection::SqliteBindValue;
 
 /// Trait for the implementation of a SQLite aggregate function
 ///
