--- diesel/src/type_impls/floats.rs
@@ -1,1 +1,0 @@
-
