--- diesel/src/type_impls/integers.rs
@@ -1,7 +1,0 @@
-use byteorder::WriteBytesExt;
-use std::error::Error;
-use std::io::prelude::*;
-
-use crate::backend::Backend;
-use crate::serialize::{self, IsNull, Output, ToSql};
-use crate::sql_types;
