+++ diesel/src/type_impls/mod.rs
@@ -1,7 +1,5 @@
 mod date_and_time;
 mod decimal;
-//pub mod floats;
-//mod integers;
 #[cfg(all(feature = "serde_json", any(feature = "postgres", feature = "mysql")))]
 mod json;
 pub mod option;
